// desk/actions/types.ts
export enum WebhooksActionTypes {
  FETCH_WEBHOOKS_REQUEST = 'FETCH_WEBHOOK_LIST_REQUEST',
  FETCH_WEBHOOKS_SUCCESS = 'FETCH_WEBHOOK_LIST_SUCCESS',
  FETCH_WEBHOOKS_FAIL = 'FETCH_WEBHOOK_LIST_FAIL',

  ADD_WEBHOOK_REQUEST = 'ADD_WEBHOOK_REQUEST',
  ADD_WEBHOOK_SUCCESS = 'ADD_WEBHOOK_SUCCESS',
  ADD_WEBHOOK_FAIL = 'ADD_WEBHOOK_FAIL',

  EDIT_WEBHOOK_REQUEST = 'EDIT_WEBHOOK_REQUEST',
  EDIT_WEBHOOK_SUCCESS = 'EDIT_WEBHOOK_SUCCESS',
  EDIT_WEBHOOK_FAIL = 'EDIT_WEBHOOK_FAIL',

  SEARCH_WEBHOOK_REQUEST = 'SEARCH_WEBHOOK_REQUEST',
  SEARCH_WEBHOOK_SUCCESS = 'SEARCH_WEBHOOK_SUCCESS',
  SEARCH_WEBHOOK_FAIL = 'SEARCH_WEBHOOK_FAIL',

  DELETE_WEBHOOK_REQUEST = 'DELETE_WEBHOOK_REQUEST',
  DELETE_WEBHOOK_SUCCESS = 'DELETE_WEBHOOK_SUCCESS',
  DELETE_WEBHOOK_FAIL = 'DELETE_WEBHOOK_FAIL',

  GET_SIGNATURE_REQUEST = 'GET_SIGNATURE_REQUEST',
  GET_SIGNATURE_SUCCESS = 'GET_SIGNATURE_SUCCESS',
  GET_SIGNATURE_FAIL = 'GET_SIGNATURE_FAIL',
}